// src/types/types.ts
export interface Category {
  id: string;
  name: string;
  name_az: string | null;
  name_ru: string | null;
  slug: string;
  description: string | null;
  image_url: string | null;
  parent_id: string | null;
  sort_order: number;
  is_active: boolean;
  created_at: string;
}